validate that the event expiration check period is less than the total retention duration
