+++ flo-server/src/server/mod.rs
@@ -1,6 +1,7 @@
 mod flo_io;
 mod channel_sender;
 mod event_loops;
+mod server_options;
 
 use futures::stream::Stream;
 use tokio_core::net::{TcpStream, TcpListener};
@@ -18,52 +19,7 @@ use std::net::{SocketAddr, Ipv4Addr, SocketAddrV4};
 use std::io;
 use engine;
 
-#[derive(Copy, Clone, PartialEq, Debug)]
-#[allow(dead_code)]
-pub enum MemoryUnit {
-    Megabyte,
-    Kilobyte,
-    Byte
-}
-
-#[derive(Copy, Clone, PartialEq, Debug)]
-pub struct MemoryLimit {
-    amount: usize,
-    unit: MemoryUnit,
-}
-
-
-impl MemoryLimit {
-    pub fn new(amount: usize, unit: MemoryUnit) -> MemoryLimit {
-        MemoryLimit {
-            amount: amount,
-            unit: unit,
-        }
-    }
-
-    pub fn as_bytes(&self) -> usize {
-        let multiplier = match self.unit {
-            MemoryUnit::Byte => 1,
-            MemoryUnit::Kilobyte => 1024,
-            MemoryUnit::Megabyte => 1024 * 1024,
-        };
-        multiplier * self.amount
-    }
-}
-
-#[derive(PartialEq, Clone)]
-pub struct ServerOptions {
-    pub port: u16,
-    pub data_dir: PathBuf,
-    pub default_namespace: String,
-    pub event_retention_duration: Duration,
-    pub event_eviction_period: Duration,
-    pub max_cached_events: usize,
-    pub max_cache_memory: MemoryLimit,
-    pub cluster_addresses: Option<Vec<SocketAddr>>,
-    pub actor_id: ActorId,
-    pub max_io_threads: Option<usize>,
-}
+pub use self::server_options::{ServerOptions, MemoryLimit, MemoryUnit};
 
 
 pub fn run(options: ServerOptions) -> io::Result<()> {
